requirements.txt
euriai==1.0.26
python-dotenv==1.1.1
pydantic==2.11.7
requests==2.32.4
pytest==8.4.1
pydantic-settings==2.9.1
streamlit==1.47.1
redis==6.2.0